+++ marshmallow/decorators.py
@@ -19,12 +19,12 @@ Example: ::
             item['email'] = item['email'].lower().strip()
             return item
 
-        @pre_load(raw=True)
+        @pre_load(pass_many=True)
         def remove_envelope(self, data, many):
             namespace = 'results' if many else 'result'
             return data[namespace]
 
-        @post_dump(raw=True):
+        @post_dump(pass_many=True):
         def add_envelope(self, data, many):
             namespace = 'results' if many else 'result'
             return {namespace: data}
@@ -38,6 +38,10 @@ Example: ::
         def validate_age(self, data):
             if data < 14:
                 raise ValidationError('Too young!')
+
+.. note::
+    These decorators only work with instance methods. Class and static
+    methods are not supported.
 
 .. warning::
     The invocation order of decorated methods of the same type is not guaranteed.
@@ -63,98 +67,75 @@ def validates(field_name):
     return tag_processor(VALIDATES, None, False, field_name=field_name)
 
 
-def validates_schema(fn=None, raw=False, pass_original=False):
+def validates_schema(fn=None, pass_many=False, pass_original=False):
     """Register a schema-level validates_schema method.
 
     By default, receives a single object at a time, regardless of whether ``many=True``
-    is passed to the `Schema`. If ``raw=True``, the raw data (which may be a collection)
+    is passed to the `Schema`. If ``pass_many=True``, the raw data (which may be a collection)
     and the value for ``many`` is passed.
 
     If ``pass_original=True``, the original data (before unmarshalling) will be passed as
     an additional argument to the method.
     """
-    return tag_processor(VALIDATES_SCHEMA, fn, raw, pass_original=pass_original)
+    return tag_processor(VALIDATES_SCHEMA, fn, pass_many, pass_original=pass_original)
 
 
-def pre_dump(fn=None, raw=False):
+def pre_dump(fn=None, pass_many=False):
     """Register a method to invoke before serializing an object. The method
     receives the object to be serialized and returns the processed object.
 
     By default, receives a single object at a time, regardless of whether ``many=True``
-    is passed to the `Schema`. If ``raw=True``, the raw data (which may be a collection)
+    is passed to the `Schema`. If ``pass_many=True``, the raw data (which may be a collection)
     and the value for ``many`` is passed.
     """
-    return tag_processor(PRE_DUMP, fn, raw)
+    return tag_processor(PRE_DUMP, fn, pass_many)
 
 
-def post_dump(fn=None, raw=False):
+def post_dump(fn=None, pass_many=False, pass_original=False):
     """Register a method to invoke after serializing an object. The method
     receives the serialized object and returns the processed object.
 
     By default, receives a single object at a time, transparently handling the ``many``
-    argument passed to the Schema. If ``raw=True``, the raw data
+    argument passed to the Schema. If ``pass_many=True``, the raw data
     (which may be a collection) and the value for ``many`` is passed.
     """
-    return tag_processor(POST_DUMP, fn, raw)
+    return tag_processor(POST_DUMP, fn, pass_many, pass_original=pass_original)
 
 
-def pre_load(fn=None, raw=False):
+def pre_load(fn=None, pass_many=False):
     """Register a method to invoke before deserializing an object. The method
     receives the data to be deserialized and returns the processed data.
 
     By default, receives a single datum at a time, transparently handling the ``many``
-    argument passed to the Schema. If ``raw=True``, the raw data
+    argument passed to the Schema. If ``pass_many=True``, the raw data
     (which may be a collection) and the value for ``many`` is passed.
     """
-    return tag_processor(PRE_LOAD, fn, raw)
+    return tag_processor(PRE_LOAD, fn, pass_many)
 
 
-def post_load(fn=None, raw=False):
+def post_load(fn=None, pass_many=False, pass_original=False):
     """Register a method to invoke after deserializing an object. The method
     receives the deserialized data and returns the processed data.
 
     By default, receives a single datum at a time, transparently handling the ``many``
-    argument passed to the Schema. If ``raw=True``, the raw data
+    argument passed to the Schema. If ``pass_many=True``, the raw data
     (which may be a collection) and the value for ``many`` is passed.
     """
-    return tag_processor(POST_LOAD, fn, raw)
+    return tag_processor(POST_LOAD, fn, pass_many, pass_original=pass_original)
 
 
-class _StaticProcessorMethod(staticmethod):
-    """Allows setting attributes on a staticmethod"""
-    pass
+def tag_processor(tag_name, fn, pass_many, **kwargs):
+    """Tags decorated processor function to be picked up later.
 
-
-class _ClassProcessorMethod(classmethod):
-    """Allows setting attributes on a classmethod"""
-    pass
-
-
-def tag_processor(tag_name, fn, raw, **kwargs):
-    """Tags decorated processor function to be picked up later
+    .. note::
+        Currently ony works with functions and instance methods. Class and
+        static methods are not supported.
 
     :return: Decorated function if supplied, else this decorator with its args
         bound.
     """
     if fn is None:  # Allow decorator to be used with no arguments
-        return lambda fn_actual: tag_processor(tag_name, fn_actual, raw, **kwargs)
-
-    # Special-case rewrapping staticmethod and classmethod, because we can't
-    # directly set attributes on those.
-    if isinstance(fn, staticmethod):
-        try:
-            unwrapped = fn.__func__
-        except AttributeError:
-            # For Python 2.6.
-            unwrapped = fn.__get__(True)
-        fn = _StaticProcessorMethod(unwrapped)
-    elif isinstance(fn, classmethod):
-        try:
-            unwrapped = fn.__func__
-        except AttributeError:
-            # For Python 2.6.
-            unwrapped = fn.__get__(True).im_func
-        fn = _ClassProcessorMethod(unwrapped)
+        return lambda fn_actual: tag_processor(tag_name, fn_actual, pass_many, **kwargs)
 
     # Set a marshmallow_tags attribute instead of wrapping in some class,
     # because I still want this to end up as a normal (unbound) method.
@@ -163,12 +144,12 @@ def tag_processor(tag_name, fn, raw, **kwargs):
     except AttributeError:
         fn.__marshmallow_tags__ = marshmallow_tags = set()
     # Also save the kwargs for the tagged function on
-    # __marshmallow_kwargs__, keyed by (<tag_name>, <raw>)
+    # __marshmallow_kwargs__, keyed by (<tag_name>, <pass_many>)
     try:
         marshmallow_kwargs = fn.__marshmallow_kwargs__
     except AttributeError:
         fn.__marshmallow_kwargs__ = marshmallow_kwargs = {}
-    marshmallow_tags.add((tag_name, raw))
-    marshmallow_kwargs[(tag_name, raw)] = kwargs
+    marshmallow_tags.add((tag_name, pass_many))
+    marshmallow_kwargs[(tag_name, pass_many)] = kwargs
 
     return fn